+++ app/src/main/java/net/gsantner/markor/dialog/CreateFolderDialog.java
@@ -8,23 +8,24 @@
 package net.gsantner.markor.dialog;
 
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.TextView;
 
 import net.gsantner.markor.R;
-import net.gsantner.markor.model.Constants;
 import net.gsantner.markor.util.AppCast;
 import net.gsantner.markor.util.AppSettings;
 
 public class CreateFolderDialog extends DialogFragment {
     public static final String FRAGMENT_TAG = "create_folder_dialog_tag";
+    // ----- KEYS -----
+    public static final String CURRENT_DIRECTORY_DIALOG_KEY = "current_dir_folder_key";
 
     private EditText folderNameEditText;
     private String currentDir;
@@ -35,13 +36,13 @@ public class CreateFolderDialog extends DialogFragment {
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
         LayoutInflater inflater = LayoutInflater.from(getActivity());
-        currentDir = getArguments().getString(Constants.CURRENT_DIRECTORY_DIALOG_KEY, "");
+        currentDir = getArguments().getString(CURRENT_DIRECTORY_DIALOG_KEY, "");
 
         View root;
         AlertDialog.Builder dialogBuilder;
 
         boolean darkTheme = AppSettings.get().isDarkThemeEnabled();
-        root = inflater.inflate(R.layout.ui__create_folder__dialog, null);
+        root = inflater.inflate(R.layout.ui__create_folder__dialog, (ViewGroup) null);
         dialogBuilder = new AlertDialog.Builder(getActivity(), darkTheme ?
                 R.style.Theme_AppCompat_Dialog : R.style.Theme_AppCompat_Light_Dialog);
 
@@ -53,20 +54,12 @@ public class CreateFolderDialog extends DialogFragment {
                 darkTheme ? R.color.dark__primary_text : R.color.light__primary_text));
 
 
-        dialogBuilder.setPositiveButton(getResources().getString(R.string.create), new
-                DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        String folder = currentDir + "/" + folderNameEditText.getText().toString();
-                        AppCast.CREATE_FOLDER.send(getActivity(), folder);
-                    }
-                });
+        dialogBuilder.setPositiveButton(getResources().getString(R.string.create), (dialog, which) -> {
+            String folder = currentDir + "/" + folderNameEditText.getText().toString();
+            AppCast.CREATE_FOLDER.send(getActivity(), folder);
+        });
 
-        dialogBuilder.setNegativeButton(getResources().getString(android.R.string.cancel), new
-                DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        dialog.dismiss();
-                    }
-                });
+        dialogBuilder.setNegativeButton(getResources().getString(android.R.string.cancel), (dialog, which) -> dialog.dismiss());
 
         AlertDialog dialog = dialogBuilder.show();
         folderNameEditText = dialog.findViewById(R.id.create_folder_dialog__folder_name);