+++ app/src/main/java/net/gsantner/markor/widget/FilesWidgetFactory.java
@@ -15,9 +15,10 @@ import android.widget.RemoteViewsService;
 
 import net.gsantner.markor.R;
 import net.gsantner.markor.model.Constants;
+import net.gsantner.markor.model.DocumentLoader;
+import net.gsantner.markor.util.ContextUtils;
 
 import java.io.File;
-import java.io.FileFilter;
 
 public class FilesWidgetFactory implements RemoteViewsService.RemoteViewsFactory {
 
@@ -29,7 +30,7 @@ public class FilesWidgetFactory implements RemoteViewsService.RemoteViewsFactory
     public FilesWidgetFactory(Context context, Intent intent) {
         _context = context;
         _appWidgetId = intent.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, AppWidgetManager.INVALID_APPWIDGET_ID);
-        _dir = new File(intent.getStringExtra(Constants.EXTRA_FOLDERPATH));
+        _dir = (File) intent.getSerializableExtra(DocumentLoader.EXTRA_PATH);
     }
 
     @Override
@@ -43,12 +44,9 @@ public class FilesWidgetFactory implements RemoteViewsService.RemoteViewsFactory
     }
 
     private void updateFiles() {
-        this._widgetFilesList = _dir.listFiles(new FileFilter() {
-            @Override
-            public boolean accept(File pathname) {
-                return !pathname.isDirectory();
-            }
-        });
+        this._widgetFilesList = _dir.listFiles(file ->
+                !file.isDirectory() && ContextUtils.get().isMaybeMarkdownFile(file)
+        );
     }
 
     @Override
@@ -64,7 +62,7 @@ public class FilesWidgetFactory implements RemoteViewsService.RemoteViewsFactory
     @Override
     public RemoteViews getViewAt(int position) {
         File file = _widgetFilesList[position];
-        Intent fillInIntent = new Intent().putExtra(Constants.NOTE_KEY, file);
+        Intent fillInIntent = new Intent().putExtra(DocumentLoader.EXTRA_PATH, file);
         RemoteViews rowView = new RemoteViews(_context.getPackageName(), R.layout.widget_file_item);
         rowView.setTextViewText(R.id.widget_note_title, Constants.MD_EXTENSION.matcher(file.getName()).replaceAll(""));
         rowView.setOnClickFillInIntent(R.id.widget_note_title, fillInIntent);